+++ packages/web-components/src/components/style-picker/components/style-picker-modules/style-picker-modules.ts
@@ -35,7 +35,7 @@ class CLABSStylePickerModules extends StylePickerModules {
    * @returns {TemplateResult<1>}
    */
   render() {
-    return stylePickerModulesTemplate();
+    return stylePickerModulesTemplate(this);
   }
 }