+++ packages/web-components/src/components/network-graph/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.15.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/network-graph@0.14.0...@carbon-labs/network-graph@0.15.0) (2025-07-09)
+
+**Note:** Version bump only for package @carbon-labs/network-graph
+
+
+
+
+
 # [0.14.0](https://github.com/carbon-design-system/carbon-labs/compare/@carbon-labs/network-graph@0.13.0...@carbon-labs/network-graph@0.14.0) (2025-05-08)